requirements.txt
Flask>=2.0
Flask-Cors>=3.0
onvif-zeep
opencv-python
numpy
sounddevice
soundfile
requests